More tests and code cleanup related to IndexSelectivity

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/api/index/sampling/IndexSamplingConfig.java
@@ -30,7 +30,7 @@ public class IndexSamplingConfig
 
     public IndexSamplingConfig( Config config )
     {
-        this.bufferSize = config.get( GraphDatabaseSettings.index_sampling_buffer_size );
+        this.bufferSize = config.get( GraphDatabaseSettings.index_sampling_buffer_size ).intValue();
         this.updateRatio = ((double) config.get( GraphDatabaseSettings.index_sampling_update_percentage )) / 100.0d;
         this.backgroundSampling = config.get( GraphDatabaseSettings.index_background_sampling_enabled );
     }
